move store progress to server cluster
